requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.13
django-embed-video==1.4.4
django-model-utils==4.2.0
django-storages==1.13.1
django-widget-tweaks==1.4.12
google==3.0.0
google-api-core==2.10.1
google-auth==2.11.1
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.0.0
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
markdown_grid_tables==0.0.1
mysqlclient==2.1.0
Pillow==9.1.0
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.1
requests==2.27.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Unidecode==1.3.4
urllib3==1.26.9
zipp==3.8.0
